Now a proper lein project
Attempts to use namespaces correctly, but the
result is something much less pleasant to
tinker with than the original solution based
around "load-file"

TODO : make it fun to tinker with again.
This is not some web backend library.
This is meant to be run in the repl.
It should be interactive and fun.